+++ app/src/main/java/com/example/spacex_kotlin/repository/model/Height.kt
@@ -9,5 +9,5 @@ data class Height(
     @Json(name = "feet")
     val feet: Double,
     @Json(name = "meters")
-    val meters: Int
+    val meters: Double
 )
