+++ app/src/main/java/com/example/spacex_kotlin/repository/model/Mass.kt
@@ -7,7 +7,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 data class Mass(
     @Json(name = "kg")
-    val kg: Int,
+    val kg: Double,
     @Json(name = "lb")
-    val lb: Int
+    val lb: Double
 )
