+++ app/src/main/java/com/example/spacex_kotlin/repository/model/PayloadWeight.kt
@@ -9,9 +9,9 @@ data class PayloadWeight(
     @Json(name = "id")
     val id: String,
     @Json(name = "kg")
-    val kg: Int,
+    val kg: Double,
     @Json(name = "lb")
-    val lb: Int,
+    val lb: Double,
     @Json(name = "name")
     val name: String
 )
